fix: idempotency

+++ setup.sh
@@ -15,9 +15,14 @@ function symlink_dotfile() {
   local full_file_path
   full_file_path="$(dotfiles_location)/$file"
 
-  mkdir -p "$(dirname "$destination")"
-  echo "Symlinking $full_file_path -> $destination"
-  ln -sf "$full_file_path" "$destination"
+  if [ ! -d "$destination" ]; then
+    mkdir -p "$(dirname "$destination")"
+  fi
+
+  if [ ! -L "$destination" ]; then
+    echo "Symlinking $full_file_path -> $destination"
+    ln -sf "$full_file_path" "$destination"
+  fi
 }
 
 function copy_dotfile() {
@@ -26,21 +31,36 @@ function copy_dotfile() {
   local full_file_path
   full_file_path="$(dotfiles_location)/$file"
   
-  # mkdir -p $(dirname $destination)
+  if [ ! -d "$destination" ]; then
+    mkdir -p "$(dirname "$destination")"
+  fi
 
-  #if [ ! -e "$destination" ]; then
+  if [ ! -d "$destination" ]; then
     echo "Copying $full_file_path -> $destination"
     cp "$full_file_path" "$destination"
-  #fi
+  fi
 }
+
+function git_clone() {
+  local repo_url=$1
+  local destination=$2
+  if [ ! -d "$destination" ]; then
+    git clone "$repo_url" "$destination"
+  fi
+}
+
+
 
 echo "setting up .zshenv"
 copy_dotfile zsh/.zshenv "$HOME"/
 
 echo "installing oh-my-zsh"
-wget https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh
-ZSH="$HOME/.config/zsh/oh-my-zsh" sh -c "$(cat install.sh) --unattended"
-rm install.sh
+if [ ! -d "$HOME/.config/zsh/oh-my-zsh" ]; then
+  wget https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh
+  ZSH="$HOME/.config/zsh/oh-my-zsh" sh -c "$(cat install.sh) --unattended"
+  rm install.sh
+  rm ~/.zshrc
+fi
 
 echo "setting up zsh"
 symlink_dotfile zsh/.zshrc "$HOME"/.config/zsh/.zshrc
@@ -48,11 +68,11 @@ symlink_dotfile zsh/.zprofile "$HOME"/.config/zsh/.zprofile
 
 echo "using powerlevel10k for zsh"
 export ZSH_CUSTOM="$HOME/.config/zsh/oh-my-zsh/custom"
-git clone https://github.com/romkatv/powerlevel10k.git "$ZSH_CUSTOM"/themes/powerlevel10k
+git_clone https://github.com/romkatv/powerlevel10k.git "$ZSH_CUSTOM"/themes/powerlevel10k
 
 echo "adding zsh plugins"
-git clone https://github.com/zsh-users/zsh-autosuggestions.git "$ZSH_CUSTOM"/plugins/zsh-autosuggestions
-git clone https://github.com/zsh-users/zsh-syntax-highlighting.git "$ZSH_CUSTOM"/plugins/zsh-syntax-highlighting
-git clone https://github.com/marlonrichert/zsh-autocomplete.git "$ZSH_CUSTOM"/plugins/zsh-autocomplete
+git_clone https://github.com/zsh-users/zsh-autosuggestions.git "$ZSH_CUSTOM"/plugins/zsh-autosuggestions
+git_clone https://github.com/zsh-users/zsh-syntax-highlighting.git "$ZSH_CUSTOM"/plugins/zsh-syntax-highlighting
+git_clone https://github.com/marlonrichert/zsh-autocomplete.git "$ZSH_CUSTOM"/plugins/zsh-autocomplete
 
 echo "finished setting up zsh... 🚀"
